Restructure Job
