+++ spb_curate/curate/api/__init__.py
@@ -1,3 +1,4 @@
 from .curate import *
 from .enums import *
+from .job import *
 from .model_diagnosis import *
